+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -5000,10 +5000,7 @@ internal constructor(
                     clientScope = clientScope,
                     userScope = userScope,
                     config = config,
-                    notificationsHandler = notificationsHandler ?: NotificationHandlerFactory.createNotificationHandler(
-                        context = appContext,
-                        notificationConfig = notificationConfig,
-                    ),
+                    notificationsHandler = notificationsHandler,
                     apiModelTransformers = apiModelTransformers,
                     fileTransformer = fileTransformer,
                     fileUploader = fileUploader,