+++ stream-chat-android-ui-guides/src/main/java/io/getstream/chat/android/guides/catalog/compose/customizingimageandvideoattachments/factory/MultimediaAttachmentFactory.kt
@@ -40,7 +40,7 @@ val customMediaAttachmentFactory = MediaAttachmentFactory(
                     .widthIn(10.dp)
                     .padding(2.dp)
                     .background(
-                        color = Color(red = 256, blue = 256, green = 256, alpha = 220),
+                        color = Color(red = 255, blue = 255, green = 255, alpha = 220),
                         shape = RoundedCornerShape(8.dp),
                     )
                     .fillMaxWidth(0.3f)
@@ -55,7 +55,7 @@ val customMediaAttachmentFactory = MediaAttachmentFactory(
                 modifier = Modifier
                     .padding(2.dp)
                     .background(
-                        color = Color(red = 256, blue = 256, green = 256, alpha = 220),
+                        color = Color(red = 255, blue = 255, green = 255, alpha = 220),
                         shape = RoundedCornerShape(8.dp),
                     )
                     .fillMaxWidth(0.35f)