+++ CHANGELOG.md
@@ -34,6 +34,9 @@
 
 ## stream-chat-android-state
 ### 🐞 Fixed
+- Fix thread replies updating channel unread count. [#6056](https://github.com/GetStream/stream-chat-android/pull/6056)
+- Fix messages from muted users increasing channel unread count. [#6061](https://github.com/GetStream/stream-chat-android/pull/6061)
+- Fix showing channels loading state when querying offline and no matching channels are found in the DB. [#6051](https://github.com/GetStream/stream-chat-android/pull/6051)
 
 ### ⬆️ Improved
 
@@ -69,6 +72,7 @@
 ### 🐞 Fixed
 
 ### ⬆️ Improved
+- Address some Lint findings [#6058](https://github.com/GetStream/stream-chat-android/pull/6058)
 
 ### ✅ Added
 
@@ -86,6 +90,23 @@
 ### ⚠️ Changed
 
 ### ❌ Removed
+
+# December 22th, 2025 - 6.30.1
+## stream-chat-android-client
+### 🐞 Fixed
+- Fix default notification channel created even if SDK notifications are disabled. [#6054](https://github.com/GetStream/stream-chat-android/pull/6054)
+
+## stream-chat-android-state
+### 🐞 Fixed
+- Fix issue causing gaps in the message list due to messages inserted out of order. [#6055](https://github.com/GetStream/stream-chat-android/pull/6055)
+
+## stream-chat-android-ui-components
+### 🐞 Fixed
+- Fix unable to open threads whose root message was deleted. [#6046](https://github.com/GetStream/stream-chat-android/pull/6046)
+
+## stream-chat-android-compose
+### 🐞 Fixed
+- Fix unable to open threads whose root message was deleted. [#6046](https://github.com/GetStream/stream-chat-android/pull/6046)
 
 # December 12th, 2025 - 6.30.0
 ## stream-chat-android-client